✨ feat: 新增 inbox 模块入口
