+++ src/pages/chat/SessionList/index.tsx
@@ -1,9 +1,11 @@
 import { DraggablePanelBody } from '@lobehub/ui';
 import { memo } from 'react';
+import { Flexbox } from 'react-layout-kit';
 
 import FolderPanel from '@/features/FolderPanel';
 
 import Header from './Header';
+import Inbox from './Inbox';
 import SessionList from './List';
 
 export const Sessions = memo(() => {
@@ -11,7 +13,13 @@ export const Sessions = memo(() => {
     <FolderPanel>
       <Header />
       <DraggablePanelBody style={{ padding: 0 }}>
-        <SessionList />
+        <Flexbox gap={8}>
+          <Inbox />
+          <Flexbox paddingInline={16} style={{ fontSize: 12 }}>
+            角色列表
+          </Flexbox>
+          <SessionList />
+        </Flexbox>
       </DraggablePanelBody>
     </FolderPanel>
   );
