+++ src/pages/welcome/Footer/index.tsx
@@ -9,6 +9,7 @@ import { CHANGELOG, DISCORD, GITHUB } from '@/const/url';
 
 const Footer = memo(() => {
   const theme = useTheme();
+
   return (
     <Flexbox align={'center'} horizontal justify={'space-between'} style={{ padding: 16 }}>
       <span style={{ color: theme.colorTextDescription }}>©{new Date().getFullYear()} LobeHub</span>
